% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaboliteIDmapping.R
\docType{package}
\name{metaboliteIDmapping}
\alias{metaboliteIDmapping}
\title{metaboliteIDmapping package}
\description{
This package loads an ID mapping table of 9 highly used
metabolite ID formats, including Pubchem, Comptox Chemical 
Dashboard, KEGG, HMDB, Drugbank, ChEbi, and CAS. 
The mapping table contains more than 1.1 million metabolite
entries.
}
\examples{
ls('package::metaboliteIDMapping')
}
\author{
Sebastian Canzler \email{sebastian.canzler@ufz.de}
}
\keyword{package}
