% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metabolitesMapping.R
\docType{data}
\name{metabolitesMapping}
\alias{metabolitesMapping}
\title{ID Mapping table of nine different metabolite ID formats}
\format{A tibble with 9 variables and over 1.1 million metabolites:
\describe{
    \item{DTXCID}{DSSTox structure identifier, character}
    \item{DTXSID}{DSSTox substance identifier, character}
    \item{CAS}{CAS registry number, character}
    \item{CID}{Pubchem compound identifier, character}
    \item{CID}{Pubchem substance identifier, character}
    \item{HMDB}{Human Metabolome Database identifier (new format), character}
    \item{ChEBI}{Chemical Entities of Biological Interest identifier, character}
    \item{KEGG}{KEGG Compound identifier, character}
    \item{Drugbank}{Drugbank identifier, character}
}}
\usage{
metabolitesMapping
}
\description{
Four different sources of annotated metabolites, i.e., \code{HMDB},
\code{ChEBI}, \code{CompTox}, and the \code{graphite} R package, 
have been retrieved to compile a comprehensive mapping of available
metabolite IDs. ID formats that are represented in the mapping table are:
DTXCID (Comptox),
DTXSID (Comptox),
CAS-number,
CID (Pubchem),
SID (Pubchem),
HMDB,
ChEBI,
KEGG,
and Drugbank
}
\examples{
metabolitesMapping
}
\keyword{datasets}
